Added something that looks like a real renderer :). It's not done yet, but I've changed a lot of things, so here's the updated code if you want to work on it. Also, osg has been replaced by another library that can handle matrices.

+++ trunk/Src/BurgerEngine/Graphics/OpenGLContext.h
@@ -29,7 +29,7 @@ public:
 	~OpenGLContext(){}
 
 	/// \brief	Initial set up of the OpenGl context
-	bool Initialize( const float a_fWindowWidth, const float a_fWindowHeight );
+	bool Initialize( const unsigned int a_fWindowWidth, const unsigned int a_fWindowHeight );
 
 	/// \brief Clear variable
 	void Terminate();
